--- cli/actors/app2/helpers.ts
@@ -1,7 +1,0 @@
-export type ActorInput<T> = {
-  input: T;
-};
-
-export type EventOutput<T> = {
-  output: T;
-};
