+++ cli/actors/helpers.ts
@@ -1,3 +1,7 @@
 export type ActorInput<T> = {
   input: T;
 };
+
+export type EventOutput<T> = {
+  output: T;
+};
