+++ cli/components/App.tsx
@@ -1,15 +1,15 @@
 import { Box } from "ink";
 import FilePicker from "./FilePicker.tsx";
-import { useApp } from "../hooks/useApp.ts";
 import Selected from "./Selected.tsx";
 import Preview from "./Preview.tsx";
 import Welcome from "./Welcome.tsx";
 import { useStdoutDimensions } from "../hooks/useStdoutDimensions.ts";
 import { Base16Provider } from "../colors/Base16Context.tsx";
 import ColorPreview from "./ColorPreview.tsx";
+import { useApp } from "../hooks/useApp.ts";
 
 export default function App({ rootDir }: { rootDir: string }) {
-  const { fileTree: context, app: appContext, preview, welcomeIsOpen } = useApp(
+  const { fileTree: context, selected, preview, welcomeIsOpen } = useApp(
     rootDir,
   );
   const { columns } = useStdoutDimensions();
@@ -20,7 +20,7 @@ export default function App({ rootDir }: { rootDir: string }) {
         {welcomeIsOpen ? <Welcome /> : (
           <Box flexDirection="column">
             <ColorPreview />
-            <Selected appContext={appContext} />
+            <Selected context={selected} />
             <FilePicker context={context} />
             <Preview preview={preview} />
           </Box>
